% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berlingske.R
\name{berlingskebarometer}
\alias{berlingskebarometer}
\title{Collects poll data from Berlingske Barometer}
\source{
\url{http://www.politiko.dk/barometeret}
}
\usage{
berlingskebarometer(year = lubridate::year(Sys.Date()))
}
\arguments{
\item{year}{is the year you want data for}
}
\value{
A data frame containing a year of Berlingske Barometer data
}
\description{
Collects poll data from Berlingske Barometer
}
\examples{
library(pollsDK)
b16 <- berlingskebarometer()
b15 <- berlingskebarometer(2015)
b14 <- berlingskebarometer(2014)
b <- rbind(b16, b15, b14)
b

}

