% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gahner.R
\name{gahner}
\alias{gahner}
\title{Collects poll data from Erik Gahners Github}
\source{
\url{https://github.com/erikgahner}
}
\usage{
gahner()
}
\value{
A data frame of all poll data from Erik Gahners Github
}
\description{
Collects poll data from Erik Gahners Github
}
\examples{
library(pollsDK)
gahner()

}

