% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gahner.R
\name{gahner_partyletters}
\alias{gahner_partyletters}
\title{Tells you which parties are in Gahners data}
\source{
\url{https://github.com/erikgahner}
}
\usage{
gahner_partyletters()
}
\value{
A list of party letters
}
\description{
Tells you which parties are in Gahners data
}
\examples{
library(pollsDK)
gahner_partyletters()

}

