% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{gahner_plot}
\alias{gahner_plot}
\title{Visualizes poll data from Erik Gahners Github}
\source{
\url{https://github.com/erikgahner}
}
\usage{
gahner_plot(pollingfirms = NULL, parties = NULL)
}
\value{
A ggplot of danish polls
}
\description{
Visualizes poll data from Erik Gahners Github
}
\examples{
library(pollsDK)
gahner_plot()

}

