% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gahner.R
\name{gahner_pollingfirms}
\alias{gahner_pollingfirms}
\title{Tells you which polling firms are in Gahners data}
\source{
\url{https://github.com/erikgahner}
}
\usage{
gahner_pollingfirms()
}
\value{
A list of polling firms
}
\description{
Tells you which polling firms are in Gahners data
}
\examples{
library(pollsDK)
gahner_pollingfirms()

}

