% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_berlingske}
\alias{get_berlingske}
\title{Collects poll data from Berlingske Barometer}
\source{
\url{http://www.politiko.dk/barometeret}
}
\usage{
get_berlingske(year = lubridate::year(Sys.Date()))
}
\arguments{
\item{year}{is the year you want data for}
}
\value{
A data frame containing a year of Berlingske Barometer data
}
\description{
Collects poll data from Berlingske Barometer
}
\examples{
library(pollsDK)
b <- get_berlingske()

}
