% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_gahner}
\alias{get_gahner}
\title{Collects poll data from Erik Gahners Github}
\source{
\url{https://github.com/erikgahner}
}
\usage{
get_gahner()
}
\value{
A data frame of all poll data from Erik Gahners Github
}
\description{
Collects poll data from Erik Gahners Github
}
\examples{
\dontrun{
library(pollsDK)
g <- get_gahner()
}

}
