% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_polls}
\alias{get_polls}
\title{Collects poll data from Berlingske}
\source{
\url{http://www.politiko.dk/barometeret}
}
\usage{
get_polls(year = lubridate::year(Sys.Date()))
}
\arguments{
\item{year}{is the year you want data for}
}
\value{
A data frame containing a year of polling data
}
\description{
Collects poll data from Berlingske
}
\examples{
\dontrun{
library(pollsDK)
p <- get_polls()
}

}
