% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_ritzau}
\alias{get_ritzau}
\title{Collects the most recent numbers from the Ritzau Index}
\source{
\url{https://www.ritzau.dk/Produkter og Services/Ritzau Index.aspx}
}
\usage{
get_ritzau()
}
\value{
A list of data frames each containing a different poll, but all containing the Ritzau Index.
}
\description{
Collects the most recent numbers from the Ritzau Index
}
\examples{
library(pollsDK)
r <- get_ritzau()

}
