% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_voter_migration}
\alias{get_voter_migration}
\title{Collect voter migration from dr.dk}
\usage{
get_voter_migration()
}
\value{
a list
}
\description{
Collect voter migration from dr.dk
}
\examples{
vote_mig <- get_voter_migration()

}
