% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{partycolors}
\alias{partycolors}
\title{Colors for the different parties}
\format{A vector with 11 colors - one for each major party.}
\usage{
partycolors
}
\description{
A vector of colors for the eleven different parties. They are sorted by their
party letter.
}
\keyword{datasets}
